if table exists for dropping bay_types

+++ db/migrate/20150413142517_create_bay_kinds.rb
@@ -1,6 +1,8 @@
 class CreateBayKinds < ActiveRecord::Migration
   def change
-    drop_table :bay_types
+    if !table_exists?("bay_types")
+      drop_table :bay_types
+    end
 
     create_table :bay_kinds do |t|
       t.string :name
